Oragnaize custom shape code
