--- editorextensions/UVexplorer-integration/src/network-device-block.ts
@@ -1,27 +1,0 @@
-import { CustomBlockProxy, EditorClient, SerializedDataError, SerializedFieldType, Viewport } from 'lucid-extension-sdk';
-
-export class NetworkDeviceBlock extends CustomBlockProxy {
-    private viewport: Viewport;
-    private library = "UVexplorer-shapes";
-    private shape = 'networkDevice';
-
-    constructor(id: string, client: EditorClient) {
-        super(id, client);
-        this.viewport = new Viewport(client);
-    }
-
-    public async createCustomBlock(company: string, deviceType: string) {
-        var customBlockDef = await this.client.getCustomShapeDefinition(this.library, this.shape);
-    
-        if (!customBlockDef) {
-            return;
-        }
-    
-        const page = this.viewport.getCurrentPage();
-        if (page != undefined) {
-            const customBlock = page.addBlock(customBlockDef);
-            customBlock.shapeData.set("make", company);  // This is how I can change the logo dynamically
-            customBlock.shapeData.set("deviceType", deviceType);
-        }
-    }
-}
